+++ src/PinguApps.Appwrite.Shared/Requests/Databases/CreateIPAttributeRequest.cs
@@ -5,6 +5,6 @@ namespace PinguApps.Appwrite.Shared.Requests.Databases;
 /// <summary>
 /// The request to create an IP Address attribute
 /// </summary>
-public class CreateIPAddressRequest : CreateStringAttributeBaseRequest<CreateIPAddressRequest, CreateIPAddressRequestValidator>
+public class CreateIPAttributeRequest : CreateStringAttributeBaseRequest<CreateIPAttributeRequest, CreateIPAttributeRequestValidator>
 {
 }
