+++ src/PinguApps.Appwrite.Shared/Requests/Databases/UpdateIPAttributeRequest.cs
@@ -5,6 +5,6 @@ namespace PinguApps.Appwrite.Shared.Requests.Databases;
 /// <summary>
 /// The requset to update an up address attribute
 /// </summary>
-public class UpdateIPAddressAttributeRequest : UpdateStringAttributeBaseRequest<UpdateIPAddressAttributeRequest, UpdateIPAddressAttributeRequestValidator>
+public class UpdateIPAttributeRequest : UpdateStringAttributeBaseRequest<UpdateIPAttributeRequest, UpdateIPAttributeRequestValidator>
 {
 }
